feat: v5 major update - new services, mobile app, and mt5 agents

+++ gas-mt5-websocket/src/models/tick.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 from typing import List, Optional
 
 class TickData(BaseModel):
-    time: int
+    time: float           # Some EAs send as float
     bid: Optional[float] = None
     ask: Optional[float] = None
     last: Optional[float] = None
